+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+Unreleased
+----------
+
+-  feat: :doc:`/scripts/sql2csv` adds a :code:`--execution-option` option.
+-  feat: :doc:`/scripts/sql2csv` uses the ``stream_results=True`` execution option, by default, to not load all data into memory at once.
+
 2.0.1 - July 12, 2024
 ---------------------